// x-pack/platform/plugins/shared/cases/server/services/user_actions/types.d.ts
export interface BuilderParameters {
    title: {
        parameters: {
            payload: {
                title: string;
            };
        };
    };
    description: {
        parameters: {
            payload: {
                description: string;
            };
        };
    };
    status: {
        parameters: {
            payload: {
                status: CaseStatuses;
                closeReason?: string;
                syncAlerts?: boolean;
                syncedAlertCount?: number;
            };
        };
    };
    severity: {
        parameters: {
            payload: {
                severity: CaseSeverity;
            };
        };
    };
    tags: {
        parameters: {
            payload: {
                tags: string[];
            };
        };
    };
    assignees: {
        parameters: {
            payload: {
                assignees: CaseAssignees;
            };
        };
    };
    pushed: {
        parameters: {
            payload: {
                externalService: PushedUserAction['payload']['externalService'];
            };
        };
    };
    settings: {
        parameters: {
            payload: {
                settings: {
                    syncAlerts?: boolean;
                    extractObservables?: boolean;
                };
            };
        };
    };
    comment: {
        parameters: {
            payload: {
                attachment: CommentUserAction['payload']['comment'];
            };
        };
    };
    connector: {
        parameters: {
            payload: {
                connector: ConnectorUserAction['payload']['connector'];
            };
        };
    };
    create_case: {
        parameters: {
            payload: CasePostRequest;
        };
    };
    delete_case: {
        parameters: {
            payload: {};
        };
    };
    category: {
        parameters: {
            payload: {
                category: string | null;
            };
        };
    };
    customFields: {
        parameters: {
            payload: {
                customFields: CaseCustomFields;
            };
        };
    };
    observables: {
        parameters: {
            payload: {
                observables: {
                    actionType: ObservablesActionType;
                    count: number;
                };
            };
        };
    };
    extended_fields: {
        parameters: {
            payload: {
                extended_fields: Record<string, string>;
            };
        };
    };
    template: {
        parameters: {
            payload: {
                template: {
                    id: string;
                    version: number;
                } | null;
            };
        };
    };
}
export interface CreateUserAction<T extends keyof BuilderParameters> {
    type: T;
    payload: BuilderParameters[T]['parameters']['payload'];
}
export type UserActionParameters<T extends keyof BuilderParameters> = BuilderParameters[T]['parameters'] & CommonArguments;
export interface CommonArguments {
    user: User;
    caseId: string;
    owner: string;
    savedObjectId?: string;
    connectorId?: string;
    action?: UserActionAction;
}
export interface Attributes {
    action: UserActionAction;
    created_at: string;
    created_by: User;
    owner: string;
    type: UserActionType;
    payload: Record<string, unknown>;
}
export interface SavedObjectParameters {
    attributes: UserActionPersistedAttributes;
    references: SavedObjectReference[];
}
export interface EventDetails {
    getMessage: (storedUserActionId?: string) => string;
    action: UserActionAction;
    descriptiveAction: string;
    savedObjectId: string;
    savedObjectType: string;
}
export interface UserActionEvent {
    parameters: SavedObjectParameters;
    eventDetails: EventDetails;
}